# docs/api/openapi.yml
openapi: 3.1.0
info:
  title: 数字小窝 API
  version: 0.1.0
  description: 宠物数字纪念网站当前可用接口。所有写操作需要同域 XSRF Cookie 与 X-XSRF-TOKEN 请求头；公开页的口令解锁也适用。
servers:
  - url: /api/v1
paths:
...
  /auth/register:
    post:
      summary: 注册账户并建立会话
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201": { description: 账户已创建 }
        "409": { description: 邮箱已注册 }
  /auth/login:
    post:
      summary: 登录并建立会话
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200": { description: 登录成功 }
        "401": { description: 凭据不正确 }
...
  /auth/me:
    get:
      summary: 获取当前账户
      responses:
        "200": { description: 当前账户 }
        "401": { description: 尚未登录 }
  /auth/export-links:
    post:
      summary: 创建一次性资料包下载链接
      description: 返回仅限当前账户、15 分钟内有效且只能消费一次的下载地址；下载内容为 ZIP，含结构化 JSON、可阅读的 Markdown、媒体清单与当前图片、短视频文件。
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: 下载地址与过期时间 }
  /auth/export-download/{token}:
    get:
      summary: 消费一次性资料包下载链接
      security: [{ sessionCookie: [] }]
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: ZIP 资料包，下载后令牌失效 }
        "410": { description: 下载链接已使用或已过期 }
  /auth/account:
    delete:
      summary: 提交当前账户的删除请求
      description: 请求体 confirmation 必须精确为 DELETE。提交后公开纪念页与图片立即下线，账户仍可在 7 天撤销期内导出资料并取消删除；到期后后台永久清理关联数据。
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeleteAccountRequest" }
      responses:
        "200": { description: 已安排删除，返回 scheduled 与 scheduledFor }
  /auth/account/deletion:
    get:
      summary: 获取当前账户的删除请求状态
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: 返回 pending 与可撤销截止时间 scheduledFor }
  /auth/account/deletion/cancel:
    post:
      summary: 取消当前账户的删除请求
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: 已取消，pending 为 false }
        "409": { description: 删除撤销期已结束，无法再取消 }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
...
  schemas:
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8, maxLength: 72 }
        displayName: { type: string, maxLength: 32 }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    DeleteAccountRequest:
      type: object
      required: [confirmation]
      properties:
        confirmation: { type: string, enum: [DELETE] }
...